Fix vertical acceleration on Linux
Attempt to fix the problem identified in: https://github.com/hrvach/deskhop/issues/274

This happens because a single 32768x32768 desktop is used for the emulated absolute mouse. The more screens present, the more distorted the vertical scaling becomes compared to the horizontal scaling.
This change multiply the vertical acceleration factor by the number of screens to keep the feel consistent.