Fix incorrect HID keyboard report parsing when Report IDs are used
The parser was incorrectly assuming that any 8-byte or 9-byte report was a
standard Boot Protocol report (Modifier, Reserved, 6 Keys). This caused
issues for devices like the Microsoft Ergonomic Keyboard that use Report IDs,
as the 1-byte Report ID was being misinterpreted as the Modifier byte.

Modified extract_kbd_data to only apply the Boot Protocol heuristic if the
interface is confirmed not to use Report IDs.